backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.7.4
python-multipart==0.0.9
aiosmtplib==3.0.1
apscheduler==3.10.4
sqlalchemy==2.0.20  # Fixed typo from sqlachemy
alembic==1.11.1
bcrypt==4.0.1  # Critical: Fixes passlib compatibility
